backend/requirements.txt
pandas>=1.3.0
matplotlib>=3.5.0
openpyxl>=3.0.0
numpy>=1.21.0
fastapi>=0.68.0
PyJWT[crypto]>=2.8.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pymongo>=4.0.0
seaborn>=0.11.0
gprofiler-official>=1.0.0
python-multipart>=0.0.5
scipy>=1.7.0
scikit-learn>=1.0.0
networkx>=2.6.0
matplotlib-venn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1